almost done with comments refactoring
